LineItemDraft to interface. #942
